+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/block/machine/blood_purifier/BloodPurifierMenu.kt
@@ -2,7 +2,6 @@ package dev.aaronhowser.mods.geneticsresequenced.block.machine.blood_purifier
 
 import dev.aaronhowser.mods.geneticsresequenced.block.base.CraftingMachineBlockEntity
 import dev.aaronhowser.mods.geneticsresequenced.block.base.menu.MachineMenu
-import dev.aaronhowser.mods.geneticsresequenced.block.base.menu.ScreenTextures
 import dev.aaronhowser.mods.geneticsresequenced.registry.ModBlocks
 import dev.aaronhowser.mods.geneticsresequenced.registry.ModMenuTypes
 import net.minecraft.network.FriendlyByteBuf
@@ -48,6 +47,14 @@ class BloodPurifierMenu(
         addDataSlots(containerData)
     }
 
+    override fun getPercentDone(): Float {
+        return if (maxProgress != 0) {
+            progress.toFloat() / maxProgress.toFloat()
+        } else {
+            0f
+        }
+    }
+
     override fun stillValid(pPlayer: Player): Boolean {
         return stillValid(
             ContainerLevelAccess.create(level, blockEntity.blockPos),
@@ -70,16 +77,6 @@ class BloodPurifierMenu(
 
     val isCrafting
         get() = progress > 0
-
-    fun getScaledProgress(): Int {
-        val progressArrowWidth = ScreenTextures.Elements.ArrowRight.Dimensions.WIDTH
-
-        return if (maxProgress == 0 || progress == 0) {
-            0
-        } else {
-            progress * progressArrowWidth / maxProgress
-        }
-    }
 
     companion object {