+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/block/machine/blood_purifier/BloodPurifierScreen.kt
@@ -14,7 +14,6 @@ class BloodPurifierScreen(
 
     override val backgroundTexture: ResourceLocation = ScreenTextures.Backgrounds.BASIC
 
-    override fun progressArrowAmountToRender(): Int = menu.getScaledProgress()
     override fun shouldRenderProgressArrow() = menu.isCrafting
 
 }