+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/block/machine/dna_decryptor/DnaDecryptorScreen.kt
@@ -13,6 +13,5 @@ class DnaDecryptorScreen(
 ) : MachineScreen<DnaDecryptorMenu>(pMenu, pPlayerInventory, pTitle) {
     override val backgroundTexture: ResourceLocation = ScreenTextures.Backgrounds.DNA_DECRYPTOR
 
-    override fun progressArrowAmountToRender(): Int = menu.getScaledProgress()
     override fun shouldRenderProgressArrow(): Boolean = menu.isCrafting
 }