+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/block/machine/incubator/IncubatorScreen.kt
@@ -34,7 +34,6 @@ class IncubatorScreen(
     override val arrowLeftPos: Int = ScreenTextures.Elements.ArrowDown.Position.X
     override val arrowTopPos: Int = ScreenTextures.Elements.ArrowDown.Position.Y
 
-    override fun progressArrowAmountToRender(): Int = menu.getScaledProgress()
     override fun shouldRenderProgressArrow(): Boolean = menu.isCrafting
 
     private var bubblePosProgress = 0