+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/block/machine/plasmid_infuser/PlasmidInfuserScreen.kt
@@ -13,6 +13,5 @@ class PlasmidInfuserScreen(
 ) : MachineScreen<PlasmidInfuserMenu>(pMenu, pPlayerInventory, pTitle) {
     override val backgroundTexture: ResourceLocation = ScreenTextures.Backgrounds.PLASMID_INFUSER
 
-    override fun progressArrowAmountToRender(): Int = menu.getScaledProgress()
     override fun shouldRenderProgressArrow(): Boolean = menu.isCrafting
 }