+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/block/machine/plasmid_injector/PlasmidInjectorMenu.kt
@@ -56,6 +56,13 @@ class PlasmidInjectorMenu(
         addDataSlots(containerData)
     }
 
+    override fun getPercentDone(): Float {
+        return if (maxProgress == 0) {
+            0.0f
+        } else {
+            progress.toFloat() / maxProgress.toFloat()
+        }
+    }
 
     override fun stillValid(pPlayer: Player): Boolean {
         return stillValid(
@@ -79,16 +86,6 @@ class PlasmidInjectorMenu(
 
     val isCrafting
         get() = progress > 0
-
-    fun getScaledProgress(): Int {
-        val progressArrowSize = ScreenTextures.Elements.ArrowRight.Dimensions.WIDTH
-
-        return if (maxProgress == 0 || progress == 0) {
-            0
-        } else {
-            progress * progressArrowSize / maxProgress
-        }
-    }
 
     companion object {
         fun showTooltip(event: ItemTooltipEvent) {