+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/block/machine/plasmid_injector/PlasmidInjectorScreen.kt
@@ -13,6 +13,5 @@ class PlasmidInjectorScreen(
 ) : MachineScreen<PlasmidInjectorMenu>(pMenu, pPlayerInventory, pTitle) {
     override val backgroundTexture: ResourceLocation = ScreenTextures.Backgrounds.PLASMID_INJECTOR
 
-    override fun progressArrowAmountToRender(): Int = menu.getScaledProgress()
     override fun shouldRenderProgressArrow(): Boolean = menu.isCrafting
 }